Small update
Fixed a bug with the kicking script causing it to be able to score multiple points. Added tutorial text to first level explaining how to jump and the object of the game. Added map6, which has similar vibes to 5 but with the kicking/soccer mechanic needed to be used. Removed sample scene 1 since it didnt add anything to the game and wasnt fun.